Update the integration test for Minecraft 1.7.2.

+++ ProtocolLib/src/test/java/com/comphenix/integration/protocol/SimpleCraftBukkitITCase.java
@@ -11,6 +11,8 @@ import java.util.Map;
 import java.util.concurrent.Callable;
 import java.util.concurrent.TimeUnit;
 
+import org.apache.logging.log4j.Logger;
+import org.apache.logging.log4j.simple.SimpleLogger;
 import org.bukkit.Bukkit;
 import org.bukkit.plugin.Plugin;
 import org.bukkit.plugin.PluginDescriptionFile;
@@ -20,6 +22,7 @@ import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.powermock.core.classloader.annotations.PowerMockIgnore;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 
 import com.comphenix.protocol.ProtocolLibrary;
@@ -29,6 +32,7 @@ import com.google.common.io.Files;
 
 // Damn final classes ...
 @RunWith(org.powermock.modules.junit4.PowerMockRunner.class)
+@PowerMockIgnore({ "org.apache.log4j.*", "org.apache.logging.*", "org.bukkit.craftbukkit.libs.jline.*" })
 @PrepareForTest(PluginDescriptionFile.class)
 public class SimpleCraftBukkitITCase {	
 	// The fake plugin
@@ -50,7 +54,14 @@ public class SimpleCraftBukkitITCase {
 	@BeforeClass
 	public static void setupCraftBukkit() throws Exception {
 		setupPlugins();
-		org.bukkit.craftbukkit.Main.main(new String[0]);
+
+		try {
+			org.bukkit.craftbukkit.Main.main(new String[0]);
+		} finally {
+			System.out.println("Current class loader: " + Thread.currentThread().getContextClassLoader());
+			System.out.println("Loader of SimpleLogger: " + SimpleLogger.class.getClassLoader());
+			System.out.println("Loader of Logger: " + Logger.class.getClassLoader());
+		}
 		
 		// We need to wait until the server object is ready
 		while (Bukkit.getServer() == null)
@@ -61,12 +72,13 @@ public class SimpleCraftBukkitITCase {
 	
 		// No need to look for updates
 		FieldUtils.writeStaticField(ProtocolLibrary.class, "UPDATES_DISABLED", Boolean.TRUE, true);
-		
+	
 		// Wait until the server and all the plugins have loaded
 		Bukkit.getScheduler().callSyncMethod(FAKE_PLUGIN, new Callable<Object>() {
 			@Override
 			public Object call() throws Exception {
 				initializePlugin(FAKE_PLUGIN);
+				ProtocolLibrary.getConfiguration().setDebug(true);
 				return null;
 			}
 		}).get(TIMEOUT_MS, TimeUnit.MILLISECONDS);
@@ -98,7 +110,9 @@ public class SimpleCraftBukkitITCase {
 		int bestLength = Integer.MAX_VALUE;
 		
 		// Copy the ProtocolLib plugin to the server
-		Files.deleteDirectoryContents(pluginDirectory);
+		if (pluginDirectory.exists()) {
+			Files.deleteDirectoryContents(pluginDirectory);
+		}
 		
 		for (File file : new File("../").listFiles()) {
 			String name = file.getName();
@@ -108,6 +122,7 @@ public class SimpleCraftBukkitITCase {
 				bestFile = file;
 			}
 		}
+		pluginDirectory.mkdirs();
 		Files.copy(bestFile, new File(pluginDirectory, bestFile.getName()));
 	}
